prime call to create api tokens
No idea what's happening here, but if you start the plugin, kill it, and let vault restart it, subsequent calls to create api tokens will fail.

The error it gives is http: ContentLength=28  with Body length 0. This is being throwing from within the net/http library and the external cause is unrelated to my code, so i have no idea what's going on. However, I discovered that if you do a primer call to the LM api first, these calls succeed every time.